[MODUSERS-384] - added userType validation
